+++ test/runtests.jl
@@ -839,6 +839,17 @@ end
     end
 end
 
+@testset "DimensionError message" begin
+  function errorstr(e)
+    b = IOBuffer()
+    Base.showerror(b,e)
+    String(b)
+  end
+  @test errorstr(DimensionError(1u"m",2)) == "DimensionError: 1 m and 2 are not dimensionally compatible."
+  @test errorstr(DimensionError(1u"m",NoDims)) == "DimensionError: 1 m and  are not dimensionally compatible."
+  @test errorstr(DimensionError(u"m",2)) == "DimensionError: m and 2 are not dimensionally compatible."
+end
+
 # Test that the @u_str macro will find units in other modules.
 module ShadowUnits
     using Unitful